FIX refactor data.ts and actions.ts functions
Added fetchData to actions.ts (imported in data.ts). It takes a string endpoint, a method and body?.
Connection status and roles are not checked anymor since API already does it.

Only uploadImage was not refactored because it needs to send a FormData body, so it is not JSON like the others and it uses res to answer, not data.
Same for getTodaysEntry, it uses res. This one could be fixed easily though. See next commits.